Cleans up lots of insert and deletion behaviors (finally).
